+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 ### Deprecated
 - Deprecated Neo blocks' `getModified()` and `setModified()` methods
 
+## 2.10.6 - 2021-07-06
+
+### Fixed
+- Fixed a bug where GraphQL queries for a Neo field's contents with the field's `level` argument set to `0` would only return level 1 blocks, rather than all blocks belonging to the field
+
 ## 2.10.5 - 2021-06-25
 
 ### Added